get cached balance
