+++ kin-ecosystem-sdk/src/main/java/com/kin/ecosystem/data/order/ExternalEarnOrderCall.java
@@ -1,15 +1,14 @@
 package com.kin.ecosystem.data.order;
 
 import android.support.annotation.NonNull;
-import com.kin.ecosystem.data.blockchain.IBlockchainSource;
+import com.kin.ecosystem.data.blockchain.BlockchainSource;
 import com.kin.ecosystem.data.order.OrderDataSource.Remote;
-import com.kin.ecosystem.network.model.Offer.OfferType;
 
 class ExternalEarnOrderCall extends CreateExternalOrderCall {
 
     ExternalEarnOrderCall(
         @NonNull Remote remote,
-        @NonNull IBlockchainSource blockchainSource,
+        @NonNull BlockchainSource blockchainSource,
         @NonNull String orderJwt,
         @NonNull ExternalOrderCallbacks externalEarnOrderCallbacks) {
         super(remote, blockchainSource, orderJwt, externalEarnOrderCallbacks);
